Bug Fixes
Fixed the highlighting issue (mostly)
Fixed the issue of clicking on a card that has no legal moves. it would move to the else statement instead of running the if statement again.
Fixed the issue of cards returning to their flipped state.

Added a flip function to fix the last issue mentioned. Updated the span when flipping fixed the issue